whisper/requirements.txt
whisperx
psycopg2-binary
python-dotenv
beautifulsoup4
ffmpeg-python
uvicorn
fastapi
webvtt-py == 0.5.1
pytest
pytest-asyncio
httpx